동현 푸쉬

+++ CarrotVillage/src/main/java/com/naver/myhome/market/Controller/FreeItemThread.java
@@ -1,16 +1,23 @@
 package com.naver.myhome.market.Controller;
 
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 import java.util.Random;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
+import org.springframework.context.annotation.ScopedProxyMode;
 
 import com.naver.myhome.market.Service.CandidateService;
 import com.naver.myhome.market.Service.UsedItemService;
 import com.naver.myhome.market.domain.Candidate;
+import com.naver.myhome.market.domain.UsedItem;
 
 @Component
+@Scope(scopeName = "request", proxyMode = ScopedProxyMode.TARGET_CLASS)
 public class FreeItemThread extends Thread {
 	private long milliSecond;
 	private int num;
@@ -24,14 +31,20 @@ public class FreeItemThread extends Thread {
 	@Override
 	public void run() {
 		try {
-			Thread.sleep(milliSecond);
-			List<Candidate> list = candidateService.selectList(num);
-			if(!list.isEmpty()) {
-				Random random = new Random();
-				int n = random.nextInt(list.size());
-				String winner = list.get(n).getId();
-				usedItemService.setWinner(winner, num);
-				candidateService.delete(num);
+			Thread.sleep(milliSecond + 1000);
+			UsedItem item = usedItemService.detail(num);
+			if(item == null) {
+				this.interrupt();
+			}
+			if(item.getWinner() == null && compareDate(item.getDeadline())) {
+				List<Candidate> list = candidateService.selectList(num);
+				if(!list.isEmpty()) {
+					Random random = new Random();
+					int n = random.nextInt(list.size());
+					String winner = list.get(n).getId();
+					usedItemService.setWinner(winner, num);
+					candidateService.delete(num);
+				}
 			}
 		} catch (InterruptedException e) {
 			e.printStackTrace();
@@ -52,5 +65,23 @@ public class FreeItemThread extends Thread {
 
 	public void setNum(int num) {
 		this.num = num;
+	}
+	
+	private boolean compareDate(String date) {
+		if(date == null) {
+			return false;
+		}
+		SimpleDateFormat format = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+		boolean resp = false;
+		try {
+			Date d = format.parse(date);
+			Date today = new Date();
+			if(d.getTime() <= today.getTime()) {
+				resp = true;
+			}
+		} catch (ParseException e) {
+			e.printStackTrace();
+		}
+		return resp;
 	}
 }
